agent/requirements.txt
livekit-agents[openai,silero,turn-detector]
python-dotenv
requests
faster-whisper